libgccjit: Add gcc_jit_context_new_array_type_u64
gcc/jit/ChangeLog:

	* docs/topics/compatibility.rst (LIBGCCJIT_ABI_37): New ABI tag.
	* docs/topics/types.rst: Document
	gcc_jit_context_new_array_type_u64.
	* jit-playback.cc (new_array_type): Change num_elements type to
	uint64_t.
	* jit-playback.h (new_array_type): Change num_elements type to
	uint64_t.
	* jit-recording.cc (recording::context::new_array_type): Change
	num_elements type to uint64_t.
	(recording::array_type::make_debug_string): Use uint64_t
	format.
	(recording::array_type::write_reproducer): Switch to
	gcc_jit_context_new_array_type_u64.
	* jit-recording.h (class array_type): Change num_elements type
	to uint64_t.
	(new_array_type): Change num_elements type to uint64_t.
	(num_elements): Change return type to uint64_t.
	* libgccjit.cc (gcc_jit_context_new_array_type_u64):
	New function.
	* libgccjit.h (gcc_jit_context_new_array_type_u64):
	New function.
	* libgccjit.exports: New function.
	* libgccjit.map: New function.

gcc/testsuite/ChangeLog:

	* jit.dg/all-non-failing-tests.h: Add test-arrays-u64.c.
	* jit.dg/test-arrays-u64.c: New test.

+++ gcc/jit/jit-playback.h
@@ -88,7 +88,7 @@ public:
   type *
   new_array_type (location *loc,
 		  type *element_type,
-		  int num_elements);
+		  uint64_t num_elements);
 
   field *
   new_field (location *loc,
